Factor out updateValidators into its own file
